requirements.txt
asttokens==2.0.5
backcall==0.2.0
black==22.6.0
click==8.1.3
cycler==0.11.0
decorator==5.1.1
executing==0.8.3
fonttools==4.34.4
graphviz==0.20
ipython==8.4.0
jedi==0.18.1
joblib==1.1.0
kiwisolver==1.4.3
matplotlib==3.5.2
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
numpy==1.23.1
packaging==21.3
pandas==1.4.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
prompt-toolkit==3.0.30
ptyprocess==0.7.0
pure-eval==0.2.2
pydotplus==2.0.2
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
sklearn-extensions==0.0.2
stack-data==0.3.0
threadpoolctl==3.1.0
tomli==2.0.1
traitlets==5.3.0
wcwidth==0.2.5